foundry fmt

+++ src/ProjectToken.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.20;
 
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
 
 contract ProjectToken is ERC20 {
     address public owner;
@@ -14,23 +14,18 @@ contract ProjectToken is ERC20 {
 
     modifier onlyApproved() {
         require(
-            msg.sender == owner ||
-                msg.sender == claimableAddr ||
-                msg.sender == tokensForProjectAddr,
-            "Only owner/approved!"
+            msg.sender == owner || msg.sender == claimableAddr || msg.sender == tokensForProjectAddr,
+            'Only owner/approved!'
         );
         _;
     }
 
-    function setClaimableAddrs(
-        address _claimableAddr,
-        address _tokensForProjectAddr
-    ) public onlyApproved {
+    function setClaimableAddrs(address _claimableAddr, address _tokensForProjectAddr) public onlyApproved {
         claimableAddr = _claimableAddr;
         tokensForProjectAddr = _tokensForProjectAddr;
     }
 
-    function mint(address to, uint256 amount) external onlyApproved {
+    function mint(address to, uint amount) external onlyApproved {
         _mint(to, amount);
     }
 }
